core/pyproject.toml
[project]
name = "oqtopus-engine-core"
version = "2.0.0"
description = "Tranqu Server: A transpiler service application using Tranqu"
readme = "README.md"
license = "Apache-2.0"
authors = [{ name = "oqtopus-team", email = "oqtopus-team@googlegroups.com" }]
keywords = []
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "License :: OSI Approved :: Apache Software License",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering",
  "Topic :: Scientific/Engineering :: Physics",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: Implementation :: CPython",
]
requires-python = ">=3.13,<4.0"
dependencies = [
  "certifi>=2025.11.12",
  "docker>=7.1.0",
  "grpcio>=1.76.0",
  "hydra-core>=1.3.2",
  "protobuf>=6.33.2",
  "pydantic>=2.12.5",
  "python-json-logger>=4.0.0",
  "qiskit==1.2",
  "qiskit-qasm3-import>=0.6.0",
  "six>=1.17.0",
  "urllib3>=2.6.1",
]

[project.urls]
documentation = "https://oqtopus-engine.readthedocs.io/"
repository = "https://github.com/oqtopus-team/oqtopus-engine"

[dependency-groups]
dev = [
  "mypy>=1.19.0",
  "pytest-asyncio>=1.3.0",
  "pytest-cov>=7.0.0",
  "pytest-mock>=3.15.1",
  "pytest>=9.0.2",
  "qiskit-aer>=0.17.2",
  "ruff>=0.14.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
preview = true
include = [
  "src/**/*.py",
  #"tests/**/*.py",
]
exclude = [
  "src/**/interfaces/**/*.py", # Generated codes
]
lint.select = ["ALL"]
lint.ignore = [
  "COM812",
  "CPY001",
  "D100",
  "D104",
  "D107",
  "D203",
  "D213",
  "ISC001",
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
  "ANN201",
  "ANN205",
  "ANN401",
  "CPY001",
  "D",
  "PLC1901",
  "PLR2004",
  "PLR6301",
  "S101",
]

[tool.mypy]
files = ["src", "tests"]
exclude = [
  "src/.*/interfaces/", # Generated codes
]

[tool.pytest.ini_options]
addopts = "--cov=src --cov-report=xml"
pythonpath = ["src"]
testpaths = ["tests"]
env = [
  "GRPC_VERBOSITY=debug",
]
